float a = (float)10/3;

+++ HelloWorld.java
@@ -4,7 +4,7 @@ public class HelloWorld {
 
 	public static void main(String[] args) {
 	
-    int a = 10/3;
+    float a = (float)10/3;
     System.out.println(a);
     }
